fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
fastapi-cache2
boto3
httpx
pytest
jsonschema
sqlalchemy
pymysql
mysql-connector-python
